storagedriver/s3: Forgoing any interface conformance tests and moved it all into S3 tests.
